Moved Image to it's own header
