+++ src/include/smol/smol_resource_manager.h
@@ -4,26 +4,12 @@
 #include <smol/smol_engine.h>
 #include <smol/smol_handle_list.h>
 #include <smol/smol_renderer_types.h>
-#include <smol/smol_font.h>
 
 namespace smol
 {
   struct Mesh;
-
-  struct SMOL_ENGINE_API Image
-  {
-    enum PixelFormat16
-    {
-      RGB_1_5_5_5     = 0,
-      RGB_5_6_5       = 1,
-    };
-
-    int width;
-    int height;
-    int bitsPerPixel;
-    PixelFormat16 format16;  // Format of 16 bit pixels
-    char* data;
-  };
+  struct Image;
+  struct Font;
 
   struct SMOL_ENGINE_API ResourceManager
   {
