+++ src/smol_renderer_gl.cpp
@@ -5,6 +5,7 @@
 #include <smol/smol_random.h>
 #include <smol/smol_renderer.h>
 #include <smol/smol_material.h>
+#include <smol/smol_image.h>
 #include <smol/smol_resource_manager.h>
 #include <smol/smol_mesh_data.h>
 #include <smol/smol_scene.h>
